# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SC Advogados",
        "description": "Visão dos endpoints para o projeto SC Advogados",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/plan": {
            "get": {
                "tags": [
                    "plan"
                ],
                "summary": "Lista todos os planos",
                "description": "Lista todos os planos",
                "operationId": "294c28d04f1cf5a756e7a8e2cfbc89df",
                "responses": {
                    "200": {
                        "description": "Lista todos os planos cadastrados"
                    }
                }
            },
            "post": {
                "tags": [
                    "plan"
                ],
                "summary": "Cadastra um novo plano",
                "operationId": "3c0ee287f14bb7c839b9d4793dc1f323",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "price_month": {
                                        "type": "number"
                                    },
                                    "price_year": {
                                        "type": "number"
                                    },
                                    "percentage_discount_questions": {
                                        "type": "number"
                                    },
                                    "percentage_discount_contracts": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "recommended": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/plan/{id}": {
            "get": {
                "tags": [
                    "plan"
                ],
                "summary": "Exibe um plano",
                "operationId": "45bfacdd2a6118139d8ebcce12e2a3b0",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "plan"
                ],
                "summary": "Remove um plano",
                "operationId": "86d9de8ed681fe5f4d3f5be01a090c25",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "api/plan/{id}": {
            "put": {
                "tags": [
                    "plan"
                ],
                "summary": "Edita uma plano",
                "operationId": "a2d6381488c2c564babe312367fe1866",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "price_month": {
                                        "type": "number"
                                    },
                                    "price_year": {
                                        "type": "number"
                                    },
                                    "percentage_discount_questions": {
                                        "type": "number"
                                    },
                                    "percentage_discount_contracts": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "recommended": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/role": {
            "get": {
                "tags": [
                    "role"
                ],
                "summary": "Lista todas as roles",
                "description": "Lista todas as roles",
                "operationId": "5382f1f8081ec0784c101b952835e6f7",
                "responses": {
                    "200": {
                        "description": "Lista todas as roles cadastradas"
                    }
                }
            },
            "put": {
                "tags": [
                    "role"
                ],
                "summary": "Edita uma role",
                "operationId": "ceda9a6ac7d304e318533788d6cb916a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "role"
                ],
                "summary": "Cadastra uma nova role",
                "operationId": "9520944bc105c406a9d07398b2ea3a17",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/role/{id}": {
            "get": {
                "tags": [
                    "role"
                ],
                "summary": "Exibe uma role",
                "operationId": "91d6146cdc7b03c517b49edc2fcba6b3",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "role"
                ],
                "summary": "Remove uma role",
                "operationId": "f6c38af5a70dd9884b06e9f1dc10cfb7",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}